Add broken implementation of recording from replay
Replays only have 30 FPS and don't store player input.
We would have to reconstruct the player input from the game state (and make it 120 FPS in the process)
or we could just move the car along rails (i.e. replaying the state instead of the inputs) until we hit the ball.
Option 1 is possibly flawed and Option 2 breaks down when hitting the ball, but might be the better one...